9077: move processDocketEntries into its own file
